pyproject.toml
[tool.poetry]
name = "cuchatai"
version = "0.1.0"
description = "Business Chatbot as client support service"
authors = ["timytapilla <ajdnmkem@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
huggingface_hub = "0.22.2"
mydependency = "0.0.0"
requests = "2.32.3"
markdownify = "0.13.1"
numpy = "1.26.4"
beautifulsoup4 = "4.12.3"
python-dotenv = "1.0.1"
pandas = "2.2.3"
unidecode = "1.3.8"
wordcloud = "1.9.3"
matplotlib = "3.9.2"
seaborn = "0.13.2"
textblob = "0.18.0.post0"
scikit-learn = "1.5.2"
networkx = "3.3"
llama-index = "0.11.14"
pinecone-client = "5.0.1"
pinecone = "5.3.1"
llama-index-core = "0.11.14"
llama-index-vector-stores-pinecone = "0.2.1"
python-frontmatter = "1.1.0"
gradio = "4.44.0"
python-telegram-bot = "^22.0"
google-api-python-client = "^2.166.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
dateparser = "^1.2.1"
pytz = "^2025.2"
flask = "^3.1.0"
